Upload data

+++ app/src/main/java/com/example/asdasdad/Fragments/DetailFragment.java
@@ -154,18 +154,23 @@ public class DetailFragment extends Fragment implements View.OnClickListener {
         deleteButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                final DatabaseReference reference = FirebaseDatabase.getInstance().getReference("recipe");
-                FirebaseStorage storage = FirebaseStorage.getInstance();
+                if(key != null){
+                    final DatabaseReference reference = FirebaseDatabase.getInstance().getReference("recipe");
+                    FirebaseStorage storage = FirebaseStorage.getInstance();
 
-                StorageReference storageReference = storage.getReferenceFromUrl(imageUrl);
-                storageReference.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
-                    @Override
-                    public void onSuccess(Void unused) {
-                        reference.child(key).removeValue();
-                        Toast.makeText(getContext(), "Delete", Toast.LENGTH_SHORT).show();
-                        Navigation.findNavController(viewF).navigate(R.id.action_detailFragment_to_show_all_recipes_fregment);
-                    }
-                });
+                    StorageReference storageReference = storage.getReferenceFromUrl(imageUrl);
+                    storageReference.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
+                        @Override
+                        public void onSuccess(Void unused) {
+                            reference.child(key).removeValue();
+                            Toast.makeText(getContext(), "Delete", Toast.LENGTH_SHORT).show();
+                            Navigation.findNavController(viewF).navigate(R.id.action_detailFragment_to_show_all_recipes_fregment);
+                        }
+                    });
+                }
+                else {
+                    Toast.makeText(getContext(), "Cant delete just remove from favorite", Toast.LENGTH_SHORT).show();
+                }
             }
         });
 
@@ -220,7 +225,7 @@ public class DetailFragment extends Fragment implements View.OnClickListener {
                     currentRecipeIsFavorite = true;
                     favoriteButton.setImageDrawable(getResources().getDrawable(R.drawable.baseline_favorite_24));
 
-                    int allEntriesSize = allEntries.size();
+                    int allEntriesSize = allEntries.size() + 1;
 
                     String jasonString = "{\"recipeName\":" + "\"" + detailName.getText().toString() + "\"" +
                             ",\"recipeImage\":" + "\"" + imageUrl + "\"" +
@@ -229,7 +234,7 @@ public class DetailFragment extends Fragment implements View.OnClickListener {
                             ",\"recipeDifficulty\":" + "\"" + detailDifficulty.getText().toString() + "\"" +
                             ",\"recipePreparationTime\":" + "\"" + detailPreparation.getText().toString() + "\"" + "}";
 
-                    editor.putString(Integer.toString(allEntriesSize + 1), jasonString);
+                    editor.putString(Integer.toString(allEntriesSize), jasonString);
                     editor.apply();
                 }
             }
@@ -259,7 +264,7 @@ public class DetailFragment extends Fragment implements View.OnClickListener {
 
     @Override
     public void onClick(View v) {
-        Toast.makeText(getContext(), "screan click", Toast.LENGTH_SHORT).show();
+        Toast.makeText(getContext(), "screen click", Toast.LENGTH_SHORT).show();
 
     }
 }
